# back-end/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Account:
    properties:
      _key:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      isActive:
        type: boolean
      lastLoginAt:
        type: string
      lastName:
        type: string
      partnerId:
        type: string
      passwordHash:
        type: string
      role:
        description: '"partner", "admin"'
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  models.AccountCreateRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      partnerId:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - partner
        - admin
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - firstName
    - lastName
    - partnerId
    - password
    - role
    - username
    type: object
  models.AccountUpdateRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      isActive:
        type: boolean
      lastName:
        type: string
    type: object
  models.ChangePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@partnerportal.com
    name: API Support
  description: API for managing partners, opportunities, products, and accounts in
    the Partner Portal
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Partner Portal API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get paginated list of all accounts with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by username
        in: query
        name: username
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Paginated accounts
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new partner account (partner must be approved)
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.AccountCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid request body
          schema:
            type: string
        "403":
          description: Partner not approved
          schema:
            type: string
        "409":
          description: Username already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Get a single account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "404":
          description: Account not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update account details (excluding password)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.AccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Account not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - accounts
  /accounts/{id}/change-password:
    post:
      consumes:
      - application/json
      description: Change password for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Current password incorrect
          schema:
            type: string
        "404":
          description: Account not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change account password
      tags:
      - accounts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"